requirements.txt
Click==7.0	
Cython==0.28.4	
GitPython==2.1.11	
# LJMPython==1.19.0	
LabJackPython==2.0.0	
Pillow==5.2.0	
PsychoPy~=2021.2.3
PyOpenGL==3.1.0
PyQt5==5.11.3
PyQt5-sip==4.19.14
PyYAML==3.13
Pygments==2.3.1
Pypubsub==4.0.0
Python-FreeType==0.6
# SMITE==0.1
SoundFile==0.10.2
XlsxWriter==1.1.5
appdirs==1.4.3
applescript==2018.11.19
arabic-reshaper==2.0.14
asn1crypto==0.24.0
astunparse==1.5.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
colorama==0.4.1
configobj==5.0.6
configparser==3.5.0
cryptography==2.3
cycler==0.10.0
decorator==4.3.0
detect==2018.11.19
dulwich==0.19.7
egi==0.9.0
esprima==4.0.1
et-xmlfile==1.0.1
freetype-py==2.0.0.post6
future==0.16.0
gevent==1.3.5
gitdb2==2.0.4
glfw==1.7.0
greenlet==0.4.14
idna==2.7
imageio==2.5.0
imageio-ffmpeg==0.2.0
ipython==7.3.0
ipython-genutils==0.2.0
jdcal==1.4
jedi==0.13.3
json-tricks==3.12.2
kiwisolver==1.0.1
lazy-import==0.2.2
lxml==4.3.2
matplotlib==2.2.2
more-itertools==7.0.0
moviepy==0.2.3.5
msgpack==0.6.1
msgpack-numpy==0.4.4.2
msgpack-python==0.5.6
numexpr==2.6.6
numpy~=1.19.5
only==2018.11.20
opencv-python==3.4.2.17
openpyxl==2.5.4
pandas==0.23.3
parso==0.3.4
pathlib==1.0.1
pathlib2==2.3.3
pickleshare==0.7.5
pip==20.2.2
pluggy==0.9.0
prompt-toolkit==2.0.9
psutil==5.4.6
psychopy-ext==0.6.0.4
psychtoolbox==3.0.15a2
psychxr==0.1.4
public==2018.11.20
py==1.8.0
pyOpenSSL==18.0.0
pyWinhook==1.6.1
pycparser==2.18
pygame~=2.1.0
pyglet==1.3.2
pylsl==1.13.1
# pyo==0.9.7
pyosf==1.0.5
pyparallel==0.2.2
pyparsing==2.2.0
pypiwin32==223
pyqmix==2018.12.13
pyserial~=3.5
pytest==4.4.0	
python-bidi==0.4.0	
python-dateutil==2.7.3	
python-gitlab==1.5.1	
python-qt5==0.1.10	
python-vlc==3.0.6108	
pytz==2018.5	
pywin32==223	
pyzmq==17.1.0	
requests==2.19.1	
ruamel.yaml==0.15.88	
runcmd==2018.12.14	
scipy==1.1.0	
setuptools==40.8.0	
sip==4.19.8	
six==1.11.0	
smmap2==2.0.4	
sounddevice==0.3.11	
tables==3.4.4	
temp==2018.11.20	
tobii-research==1.7.0	
tornado==6.0.1	
tqdm==4.23.4	
traitlets==4.3.2	
urllib3==1.23	
# vlc==2018.11.20	
wcwidth==0.1.7	
wheel==0.31.1	
winpython==1.0	
wxPython==4.0.4	
xlrd==1.1.0	
zmq==0.0.0	
pylink~=0.3.3